chore(local): move types
